requirements.txt
flask
werkzeug
torch
numpy
Pillow
ultralytics
opencv-python